lint and build fix
